+++ examples/polyscope-examples/tangency-explorer.cpp
@@ -451,9 +451,9 @@ void myCallback()
     // Only authorize selection on the input surface and the reconstruction
     auto surf = polyscope::getSurfaceMesh("Input surface");
     goodSelection = goodSelection || (selectedSurface == surf);
-    const auto nv = selectedSurface->nVertices(); 
-    const auto nf = selectedSurface->nFaces(); 
-    const auto ne = selectedSurface->nEdges(); 
+    const auto nv = surfmesh.nbVertices();
+    const auto nf = surfmesh.nbFaces();
+    const auto ne = surfmesh.nbEdges();
     // Validate that it its a face index
     if ( goodSelection )
       {
@@ -472,24 +472,11 @@ void myCallback()
             face_idx  = idx - nv;
             is_selected  = true;
             selected_kpoint = Point::zero;
-            for ( auto i : selectedSurface->faces[ face_idx ] )
+            for ( auto i : surfmesh.incidentVertices( face_idx ) )
               selected_kpoint += digital_points[ i ];
             selected_kpoint /= 2;
             std::ostringstream otext;
             otext << "Selected face = " << face_idx
-                  << " pos=" << selected_kpoint; 
-            ImGui::Text( "%s", otext.str().c_str() );
-          }
-        else if ( idx - nv - nf < ne )
-          {
-            edge_idx  = idx - nv - nf;
-            is_selected  = false; // true; // ne fonctionne pas
-            selected_kpoint = Point::zero;
-            for ( auto i : selectedSurface->edgeIndices[ edge_idx ] )
-              selected_kpoint += digital_points[ i ];
-            selected_kpoint /= 2;
-            std::ostringstream otext;
-            otext << "Selected edge = " << edge_idx
                   << " pos=" << selected_kpoint; 
             ImGui::Text( "%s", otext.str().c_str() );
           }